[#296] chore: 테스트더블 파일 분리
