--- Damago/DamagoRepositoryTests/DailyQuestionRepositoryTests/DailyQuestionRepositoryTestDoubles.swift
@@ -1,128 +1,0 @@
-//
-//  DailyQuestionRepositoryTestDoubles.swift
-//  DamagoRepositoryTests
-//
-//  Created by Eden Landelyse on 2/4/26.
-//
-
-import Foundation
-import Combine
-@testable import Damago
-@testable import DamagoNetwork
-
-final class MockNetworkProvider: NetworkProvider, @unchecked Sendable {
-    var requestHandler: ((EndPoint) async throws -> Any)?
-    var requestSuccessResult: Bool = true
-    var requestSuccessError: Error?
-    
-    var requestCalledWith: EndPoint?
-
-    func request<T: Decodable>(_ endpoint: EndPoint) async throws -> T {
-        requestCalledWith = endpoint
-        if let handler = requestHandler {
-            return try await handler(endpoint) as! T
-        }
-        fatalError("MockNetworkProvider: requestHandler not set")
-    }
-
-    func requestString(_ endpoint: EndPoint) async throws -> String {
-        requestCalledWith = endpoint
-        return ""
-    }
-
-    func requestSuccess(_ endpoint: EndPoint) async throws -> Bool {
-        requestCalledWith = endpoint
-        if let error = requestSuccessError { throw error }
-        return requestSuccessResult
-    }
-}
-
-final class MockTokenProvider: TokenProvider, @unchecked Sendable {
-    var idTokenHandler: (() async throws -> String)?
-
-    func idToken() async throws -> String {
-        return try await idTokenHandler?() ?? "mock_id_token"
-    }
-
-    func fcmToken() async throws -> String {
-        return "mock_fcm_token"
-    }
-}
-
-final class MockFirestoreService: FirestoreService, @unchecked Sendable {
-    var observeDataHandler: ((String, String) -> AnyPublisher<Result<Data, Error>, Never>)?
-    var observeCalledWith: (collection: String, document: String)?
-    
-    func observe<T: Decodable>(
-        collection: String,
-        document: String
-    ) -> AnyPublisher<Result<T, Error>, Never> {
-        observeCalledWith = (collection, document)
-        
-        if let handler = observeDataHandler {
-            return handler(collection, document).map { result in
-                switch result {
-                case .success(let data):
-                    do {
-                        let decoded = try JSONDecoder().decode(T.self, from: data)
-                        return .success(decoded)
-                    } catch {
-                        return .failure(error)
-                    }
-                case .failure(let error):
-                    return .failure(error)
-                }
-            }.eraseToAnyPublisher()
-        }
-        return Empty().eraseToAnyPublisher()
-    }
-
-    func observeQuery<T: Decodable>(
-        collection: String,
-        field: String,
-        value: Any
-    ) -> AnyPublisher<Result<[T], Error>, Never> {
-        return Empty().eraseToAnyPublisher()
-    }
-}
-
-@MainActor
-final class MockDailyQuestionLocalDataSource: DailyQuestionLocalDataSourceProtocol, @unchecked Sendable {
-    var fetchQuestionResult: DailyQuestionEntity?
-    var fetchQuestionError: Error?
-    
-    var fetchLatestQuestionHandler: (() async throws -> DailyQuestionEntity?)?
-    var saveQuestionCalledWith: DailyQuestionEntity?
-    
-    var updateAnswerCalledWith: (questionID: String, user1Answer: String?, user2Answer: String?, bothAnswered: Bool, lastAnsweredAt: Date?)?
-    var updateAnswerError: Error?
-    
-    func fetchQuestion(id: String) async throws -> DailyQuestionEntity? {
-        if let error = fetchQuestionError { throw error }
-        return fetchQuestionResult
-    }
-
-    func fetchLatestQuestion() async throws -> DailyQuestionEntity? {
-        if let error = fetchQuestionError { throw error } // reuse error for simplicity
-        return try await fetchLatestQuestionHandler?()
-    }
-
-    func saveQuestion(_ entity: DailyQuestionEntity) async throws {
-        saveQuestionCalledWith = entity
-    }
-
-    func updateAnswer(
-        questionID: String,
-        user1Answer: String?,
-        user2Answer: String?,
-        bothAnswered: Bool,
-        lastAnsweredAt: Date?
-    ) async throws {
-        if let error = updateAnswerError { throw error }
-        updateAnswerCalledWith = (questionID, user1Answer, user2Answer, bothAnswered, lastAnsweredAt)
-    }
-
-    func saveDraftAnswer(questionID: String, draftAnswer: String?) async throws {}
-    func loadDraftAnswer(questionID: String) async throws -> String? { nil }
-    func deleteDraftAnswer(questionID: String) async throws { }
-}
